hw5 rework, move complicated solution to internal/unused, write simple one in run.go
